backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
chromadb==0.4.22
openai>=1.40.0
python-dotenv==1.0.0
pytest==7.4.3
